Add <rdar://23719142>
This radar is specific around the behavior of trying to use:

```swift
let foo: String? = map.optionalFrom("a") ?? map.optionalFrom("b")
```

Where the first of these would be inferred (correctly) as
`Optional<String>`, but the casting would fail. This seems to be caused
by there being very similar function signatures with `optionalFrom`. For
example the default `optionalFrom` and then `optionalFrom` with
`RawRepresentable`