Make transformations throw MissingFieldErrors
When implementing custom errors I decided to make the transformation
function swallow `MissingFieldError`s since we were using `rethrows` and
it would have been a breaking API change. Now we're going to bump for a
breaking API change, and this is more true to what is actually going
wrong in order for the function to throw.