Remove force trys from tests
When working on large changes to Mapper, these force trys and force
casts would often end up with annoying test failures. Instead now we
handle the optionals, or the error handling as needed. We also lean a
little bit on the fact that we can remove `throws` from initializers of
`Mappable` and it still works since that isn't part of the function
signature.