Replace valueForKeyPath with custom function
Currently `valueForKeyPath:` is a little more generic than we need it to
be, and depending on the scenario this can cause an `NSException` to be
thrown, which cannot be caught from Swift.

Here is a case that will fail:

```swift
let dictionary = ["user": "not dictionary"] as NSDictionary
dictionary.valueForKeyPath("user.phone") // Crashes
```

But really in our use case, we don't care about non-dictionary objects
to be key value coding compliant for the key we're interested in.
Luckily, if the key is nil, it will not crash:

```swift
let dictionary = [:] as NSDictionary
dictionary.valueForKeyPath("user.phone") // Returns nil
```

To fix this, since `valueForKey` also crashes in these cases, we will
instead just treat everything as an `NSDictionary` and fetch the keys
accordingly.