Fix Float mapping with downcasting (#142)
The behavior of `NSNumber as? Float` in Swift is such that some values
that cannot "safely express the value stored" in NSNumber. This means
only using `as?` to cast to `Float` can fail, where doing `.floatValue`
or `Float.init` would succeed. To work around this oddity, this new
`Convertible` conformance calls `.floatValue` on `NSNumber`, leading to
a float, that is possibly less precise than you'd like.